moves locales selector logic into file
