+++ modules/asset/ui/src/index.js
@@ -1,16 +1,5 @@
+import localesSelector from './_locales';
+
 export default () => {
-  const localesSelectorElem = document.querySelector('[data-locales-selector]');
-  const toggleBtn = localesSelectorElem.querySelector('[data-locales-toggle]');
-  const localesList = localesSelectorElem.querySelector('.locales__list');
-
-  window.addEventListener('click', ({ target }) => {
-    if (localesList.classList.contains('active') && !localesSelectorElem.contains(target)) {
-      localesList.classList.remove('active');
-    }
-  });
-
-  toggleBtn.addEventListener('click', () => {
-    const shouldRemove = localesList.classList.contains('active');
-    localesList.classList[shouldRemove ? 'remove' : 'add']('active');
-  });
+  localesSelector();
 };
